requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake
apache-airflow-providers-microsoft-azure
requests
pandas
azure-storage-blob
snowflake-connector-python
azure-core 
azure-common 
